FUN-5093  Redevelop Novell Netware early binding DLS solution
Changed getUserKeys method so that if there is no  user principal it simply returns with an empty list of keys instead of deliberately throwing a null pointer exception.   A user request may not have a principal if it has not been authenticated  so it would be more appropriate for it to fail with an unauthorized message later (due to no user keys) rather than a thrown exception which implies an internal error.

git-svn-id: https://svn.cbr.au.funnelback.com/svn/enterprise-search/trunk/java/publicui@31839 3cdb6afb-a121-0410-a22c-8da0070c90a1