[FUN-5436] Further work on search history, trying to extract facet constraints on the UI side to display them next to each previous query.
In order to support both facets constraints coming from "f.Facet|X=..." and "facetScope..." I had to review the encoding of such params. The encoding of the faceted navigation parameters *names* was wrong, as they weren't encoded (but the values were), resulting in non-encoded pipes in the URL.

Both ways to transmit constraints are now encoded similarly, albeit not exactly identical: For some reasons one uses '+' for spaces, and the other is using '%20'. This shouldn't matter in practice if we compare the *decoded* forms to decide if the search we're about to save has already been run or not (TBD).



git-svn-id: https://svn.cbr.au.funnelback.com/svn/enterprise-search/trunk/java/publicui@36008 3cdb6afb-a121-0410-a22c-8da0070c90a1