FUN-6349 Add new WriterFile class to common, which encapsulates a PrintWriter and the File object it is writing to. This is needed because no obvious Java Writer object provides a getFile() method, even if the Writer was created to write to a file on disk.
SharedLogsContainer now creates these objects in a setup method, to provide access to all of the web crawler's
shared log files from a single place. A getWriterFiles() method then uses Java reflection to return a list of 
all of these objects.

This list is used in a close() method to close all of the streams if the crawler is halted or shuts down naturally.

The list is also used to generate a Map of filenames and their sizes at the time of the method call (e.g. at
a checkpoint). A persistLogSizes() method then converts this object to a JSON string and writes it to file. 

A resetLogs() call can then read in the JSON, convert back to a Map and reset (truncate) all the files to 
their previous size.

Aside: Remove WebPushStore (crawler URLStore interface to old Push system) - not currently used by anyone.

git-svn-id: https://svn.cbr.au.funnelback.com/svn/enterprise-search/trunk/java/publicui@38156 3cdb6afb-a121-0410-a22c-8da0070c90a1