Shifted the Groovy Trigger and Action implementations from publicui-web to publicui-core. Also removed the custom resource manager system I added in the last commit in favour of having the GroovyTrigger and GroovyAction manage the loading and refreshing of the Groovy classes directly. That turned out to be a lot less code overall (and if need be we can reintroduce it if/when the complexity turns otu to be warranted).
Still need to decide where we can move the CuratorYamlConfigResource and CuratorJsonConfigResource classes to so the admin-api can use them - That's not obvious because they depend on both the springmvc-common resource manager and the publicui-core curator triggers and actions. Possibly the best solution will be to duplicate them, but I'll have to investigate further.


git-svn-id: https://svn.cbr.au.funnelback.com/svn/enterprise-search/trunk/java/publicui@39395 3cdb6afb-a121-0410-a22c-8da0070c90a1