Prevent empty `form` parameters to be passed in
Be consistent with the `profile` parameter: Fall back to a default value
if it's not set, but do not allow an explicitly blank value. To do so:

* Tighten the validation regex on the field
* Remove the editor to assign a default value if not set

I've updated the release notes with that change.

Aside: Fix unit test with `CollectionEditor` now that we try to parse
the `text` as a `CollectionId`.