Override equals so objects can be compared.
In the admin API to add support for working out if a item has been modifed between LIVE and PREVIEW profile views we require that equals compares the fields of the curator models rather than using Objects default implementation.

Also moved code for listing curator Action and triggers here so that we can test on a basic level that equals has been overriden.