Converted CORS Interceptor to a Servlet Filter
An interceptor will only run when a controller mapping is found. Problem
is that CORS sometimes requires an `OPTIONS` pre-flight request and our
controllers are not mapped to the `OPTIONS` method.

A servlet filter in the other hand will run for any type of request so
when an `OPTIONS` pre-flight request is received, it will correctly add
the CORS headers in the response, allowing the browser to proceed with
the actual `GET` request.