[FUN-8076] Mitigate search session problem on extra searches
Search session features (history + cart) doesn't work on extra searches
because they're run in a separate thread but the database transaction is
thread bound. By the time we reach the view (FreeMarker, JSON, ...) the
transaction has been closed (because the extra search completed) and
lazily initialised collections cannot be fetched from the DB.

Mitigate the issue by fetching the session data only on the main search,
but not on any extra search type.