Fix click tracking links following recent Curator changes
The previous change caused 2 issues:
* The link query string would start with `?&` (works, but ugly)
* `rank` was missing a `&` prefix so it would get appended to the
previous parameter value, resulting in links having parameters like
`&profile=_defaultrank=8`

The unit test didn't detect those problems because it was using
substring match (so there was a string with `rank=8`). Changed to
actually parse the query string and work with parameters.

While doing so, found and fixed a bug with `QueryStringUtils` when parameters
contains equal signs.