[FUN-8617] Change the default output format for completions
The `simple` format has been deprecated for some time. It was changed in
the default FTL file to use `extended`, but it wasn't changed in the
default config of the jQuery plugin.

jQuery plugin is now using `extended` as well for consistency (and to
cater for cases where the form doesn't force the `format` parameter,
although I don't think it ever happened)