[SUPPORT-2099] Set the appropriate content-type when JSONP is used
When JSONP is used (by passing a `callback` parameter to the suggestion
endpoint) the content-type of the response must be set to
`application/javascript`. Browsers will refuse to execute the JSONP
callback if the content-type is different

Updated controller to override the content-type is `callback` is set,
otherwise use the default value configured for the view (defined in the
Spring XML).