[FUN-8725] Fix extra searches logging + logging improvements
Extra searches are run in different threads. The logging context is set
by a web interceptor, but extra searches don't go through an HTTP
request lifecycle because they're triggered directly in the code.
Because of that their logging context (collection, profile) may be
missing or wrong.

To prevent this, force the collection and profiles a few layers down
when we process the search transaction, so that it will work identically
between the main search and extra searches. It also has the benefit that
`profile` will always be set as we read it from the `SearchTransaction`
rather than just from the URL parameter (the profile parameter may be
forced in a hook script after we read it from the URL).

The logging context is not cleaned at the end of the search transaction
because we want it to persist for the rest of the code execution such
as processing the FTL template or serialising to JSON / XML.

Also added a few logging improvements:
* Stop FreeMarker from logging its exceptions directly. Because both
FreeMarker and Jetty are using Java Util Logging the FreeMarker
exception were ending up in the Jetty log causing a lot of noise
* Added a servlet filter to catch unhandled exceptions. Some exceptions
can't be caught by exception handler defined at the controller level
because they are thrown after the controller has run (such as FreeMarker
failures to parse a template)