[FUN-8883] Fix web resource paths containing spaces
Support space (and other odd characters) in the web resource file names
by decoding the path rather than taking the bare string we got from the
URI. That's required as we can't use Spring `@PathVariable` as the file
path can contain a slash and `@PathVariable`s only work *within* path
segments.

Updated the class to improve testability and avoid relying on the Spring
resource handler to throw an `IllegalStateException` to confirm that it
was called. Removed the test files in `dummy-search_home` since they're
not needed any more as things are mocked.

Using mocks revealed a bug in the path validation logic: we were using
`.matches()` instead of `.find()` for the validation regex so it would
only match if the path was *exactly* `..` or `/` or `\` or `:`. Use
`.find()` instead, but we then need to remove `/` from the validation
regex as the path is permitted to contain slashes (for subfolders),
instead check that the path does not start with a slash.

In practice I don't think the validation is necessary anyway because the
Spring resource handler should now allow paths that are outside the
permitted `locations` we pass to it.