[FUN-8677] Added a controller to purge old session data
Use `EntityManager.remove()` rather than using a `DELETE` query to make
sure the deletion is cascaded to associated tables (tables holding the
metadata field values).

Access to the purge URL is restricted to `localhost` only to prevent
arbitrary users to delete sessions remotely. That should be suitable to
be used in a cron job with `curl` for instance.