[FUN-9200] Fix error handling (overall + display url processor specific)
The search transaction processor wasn't handling errors other than the
one it knew about (input, output, data fetching exceptions), leaving the
possibility of a processor throwing an exception that would bubble up.

As a result the `error` field of the result packet was never set, and
since the search controller sets the HTTP response status code depending
on this field the status code would stay at `200`

Additionally, catch errors directly in `FixDisplayUrls` and output a
warning if a URL cannot be processed. That may cause issues further down
but it's already the case because when a transaction goes in error we
already leave the packet as-is (partially processed). The responsibility
of checking for errors and handling them is to the template author, I
guess we could change that if we think that's better but to be done
separately.