Fix issue with search/click history and non-search endpoints
AA and CA endpoint were passing a `SearchUser` to the
`SearchController`, but because they did not subclass
`SessionController` the user they obtained (via `@ModelAttributes`) was
not populated with a UUID. It was just the result of Spring calling the
default constructor on `SearchUser`, whereas `SessionController`
populates it with the correct UUID (either taken from a directly-managed
Cookie or from the J2EE session).

That resulted in Hibernate later trying to insert an entry in the
Session database with a user with a NULL id, and failing. I saw the
warnings while running the app. on Tomcat/Eclipse because they're logged to
stderr for some reason. I'm not sure where they end up in production as
I couldn't find them neither in `$SEARCH_HOME/log/` or
`$SEARCH_HOME/web/logs/`.

To fix it, just pass `null` for the user as we don't want to record any
kind of search history for CA & AA requests. Also removed the unused user
parameter from the recommender controller.

Aside: Disabled query logging on accessibility auditor requests.