Add a `selected` boolean attribute to facet category values
This provide a new boolean attribute `selected` on each value returned
for the facets, indicating if the value is currently selected or not.

This will make things easier for clients (templates, JSON consumers) as
they won't need to work out if a facet is currently selected by looking
at multiple nodes in the data model (faceted navigation config, question
selected category values, etc.).

I think it will be especially useful for multi-checkbox facets, but also
in the general case and will make things simpler in the new
accessibility auditor code (which was the primary motivation)