[FUN-9043] Allow facet categories to set dynamic query processor options
Update the facet category definition interface with a method allowing
categories definition to produce QPOs depending on the search question.
That allows dynamic options to be set, such as `-count_urls` with the
relevant value depending on a currently selected category value (e.g. if
the current selected value URL has 4 segments, we set `-count_urls` to
4+1).

Added an input processor to call this new method for all facet
categories, separate from the existing faceted navigation input
processor for readability.

Removed `qpOtions` attribute from the faceted navigation config and
associated code, since the QPOs are now built dynamically.

Different categories can produce similar or identical QPOs, for instance
`-rmcf=[a]` and `-rmcf=[b]`. These need to be merged in a single list
`-rmcf=[a,b]`. This is done by a new set of classes implementing the
`QueryProcessorOptionReducer` interface. Different strategies have been
implemented such as separating values with commas, or only retaining the
greatest numeric value, etc.

This will be re-usable if we also want to merge user supplied query
processor options later.

Aside: Fix issue with H2 for unit tests, since the upgrade to the latest
version it doesn't support relative paths to DB file anymore, so use an
in-memory DB instead.