[FUN-9182] Tweak AA QPOs & Facets
Tweak facets & QPOs as needed by the front end, and also to account for
bug fixes in PADRE for the group / sum options.

Force sorting for now to be able to return the top documents with the
most occurrences in the overview report. Will probably need to be
reviewed for the other screens...