[FUN-9043] Fix `-count_urls` value for URL fill facets
PADRE `-count_urls` behaves a bit non-intuitively, causing issues on the
initial implementation of dynamic QPOs for facets as I misunderstood it.

* `-count_urls=0` defaults to `5`
* `-count_urls=1` will return an actually invalid URL, in the form of
`http://example.org` (note the missing trailing slash)
* The for each subsequent value, it counts the path segments but always
omit the trailing slash, so `-count_urls=2` may return
`http://example.org/folder` without the trailing slash, making it
impossible to find out if `folder` is a folder or an actual HTML page.

As a result when computing the dynamic value of `-count_urls` we must
ensure that:
* We start at 1
* We always increment by `number of current path segments + 2`

With that approach, if the current scope is `http://example.org/folder/`
and we want to know which sub-folders are available, we need to set it
to 1 (start value) + 1 (current number of paths segments) + 2
(increment) = 4. That will return
`http://example.org/folder/sub-folder/something` so that we can confirm that
`sub-folder` is actually a folder (but we can't know if `something` is).

I think it would be significantly less confusing to fix it in PADRE so
that it always puts trailing slashes when the URL points to a folder,
but I'm not sure how difficult that is.