[FUN-9338] Fix issue with non-ascii characters when appending to XML logs
`RandomAccessFile.writeByte(String s)` does according to the Javadoc:

> Each character in the string is written out, in sequence, by
> discarding its high eight bits

I don't know why anybody would want that, but that's clearly wrong for
us as we need to write the data as UTF-8. Updated code to write a byte
array instead and get the byte array from the String in UTF-8 form.

Added corresponding unit test.