First version where SAML authentication via spring security 'works'
Much is left to do, and it may be wiser to implement it fresh than use this
as a starting point, but it's enough to prove that its possible to do it, and
may be helpful as a reference.

Points still needing investigation
* Is it practical to remove squizapi now (it forces us to override its strange version of not-yet-commons-ssl)
* How do we expose configuration of SAML
** In particular, what do we need to expose for the keystore, how do we specify certificates to trust etc
* How (in the admin world) do we map SAML users to Funnelback permissions
* Can we use SAML authentication alongside other options or...
** In particular, FACS also want a straight X.509 authentication
* Can we make everything SAML (i.e. ship with an IdP doing our existing auth) for consistency
** Would SAML be ok on its own, or do we need other options for API calls for example
** Can we make the html for marketing-dashboard and implementer-dashboard secure (avoiding them needing to know about login)

And no doubt other things I haven't thought of yet.

Still, glad to have this sort of cobbled together at least ;)