Fix FreeMarker Doco module to build in the correct order
Because the FreeMarker module didn't declare an explicit dependency
against the FTL Templates project, the Maven Reactor was building it in
the wrong order. That usually worked because the FTL templates were
already deployed so they could be fetched from the Maven repo. However
when doing the weekly version update it would fail.

So in practice it seems declaring the dependency against the FTL
templates in the dependency plugin is not good enough, it needs to be an
explicit dependency of the project, which kind of makes sense.

Replaced `unpack` with `unpack-dependencies` now that it's an explicit
dependency. Also removed extra configuration to use the default unpack
location.

Aside: Renamed folder to be consistent wit the artifactId, that got me
confused for a while.

Aside: Added name to the project to be consistent with other Modern UI
projects and make it clearer when looking at the reactor order (esp.
when running the whole of java-reactor during a weekly version update
and looking at a long list of projects).