[FUN-9182] Added processor to populate AA facet labels
AA facets have technical values like `1` for the `Perceivable`
principle, or `1.3.1` for the `Info and Relationships` success
criterion. In order to display human readable labels on the front-end,
added an output processor to populate the labels based on the WCAG
schema.

I thought about doing it client side, but I would have taken the labels
from the WCAG schema that I would have got from an Admin API call
anyway, so that would have been the same. Doing it server side is also a
bit simpler than in Angular.

We'll have to consider i18n at some point, but it will have to be done
on the AA model classes anyway so there's no real benefit in doing it
client side I think.