Stop returning 403 on content optimiser
Collection level config does not have jetty/urls ports which meant that all requests where blocked. Switching to server config makes this work again. I changed to using jetty admin port rather than the admin URLs port, as I think the port of the connection to jetty is important rather than the port of the connection to the reverse proxy.