[FUN-9638] Account for edge case with deep sub-folders
When there is only one sub-folder at the root, second and third levels
the URL scope fill facet doesn't return any result. For instance:

http://example.org/folder/sub-folder/sub-subfolder/...

In this case the URL facet will be empty because PADRE URL counts don't
go deep enough to list the `sub-subfolder` level and its siblings. All
other URL counts have only 1 entry (the folder themselve, e.g. `folder`
and `sub-folder`) so they're not returned by the facet (as this facet
type attempt to prevent presenting results with only 1 value)

To account for that, when we can't get a URL count with more than one
entry we'll retain the "deepest" one (the one with most segments) and
display it. That will allow the user to drill down to it, causing PADRE
to return new URL counts with the constraint applied, showing the
siblings folders.

Aside: Also fix an issue when the URL already ends with a trailing
slash, don't add an extra one