Add force Mime type filters.
Sometimes the client gives us a CSV,XML,JSON file that does not have the correct MIME type. Often for CSV, JSON or XML collections they consist of only gathering one type of document so these filters can be used to set the mime type which allows other filters to run. For example when gathering a csv document with the crawler at a URL which had a path that ended in .txt and returned no content the ForceCSVMime could be used before the CSVToXML filter which resulted in me being able to crawl CSV in a very small amount of time.