Creates an end-point for getting the acknowledgment counts.
FUN-9662

In the indexed term count section the id of each acknowledgment that was applied and the count of how many times it was applied will be shown.

Unlike most modern UI endpoints this is conscious of speed, in this we override all values that are about generating more data while not modifying the result set. For example rmcf is set to run over no fields. Why is this done? the countIndexedTerms option adds 77ms to count all authors for fed-gov-dave, if we run rmcf over author for fed-gove-dav it adds 770ms. If we had faceting set up over 10 metadata classes we could blow out the time to 7700ms. We of course know we don't need to do this as rmcf does not affect the results of countIndexedTerms and so it is turned off.