Add the query string as a map in the question
This is needed for hook scripts (esp. Stencils) to be able to manipulate
and construct URLs, for example when they need to alter a single
parameter but retain all the others.

Having it as a map allows them to clone the map, make the relevant
changes to it, and serialize it as a `String` using
`QueryStringUtils.toString()`.

This will also be needed when we modify the data model to generate URLs
in the Java backend rather than FreeMarker, as there's currently now way
for the input / output processors to access the current request / URL.