Change approach slightly to use an immutable map
Without an immutable one there's a chance users will update map entries
without realising it (i.e. adding an additional value to an existing
parameter value list like facets).

Using an immutable map makes that impossible, with the drawback that we
now need a setter to sets the map value when the question is
initialized, but I think there's less chance of messing things up that
way as calling to `setQueryStringMap()` is fairly explicit.