Revised query string map approach again to return a copy of the map
We in fact want users to be able to easily manipulate the map to
add/remove/update query string parameters. So instead of returning an
immutable map that would force them to manually copy it, replaced it
with a convenience method to get a mutable copy of the map that can be
used as-is.