Support multiple values for the `s` parameter
When building the query string passed to PADRE, only the first value of
`s` would be considered (as it's read via `inputParameterMap`, which is
a view on the first value of every entry in `rawInputParameters`).

This is a bit inconvenient as it means we can't populate `s` with an
array inside a hook script, but have to use a string instead:

```
transaction.question.rawInputParameters["s"] = ["subject:ACTG", "number:301"]
```

vs.

```
transaction.question.rawInputParameters["s"] = "subject:ACTG number:301"
```

...which has a few drawbacks such as having to implement string
concatenation rules if `s` is updated in multiple places (inside a
single hook script, or multiple ones), such as testing if it already has
a value or not, if it ends in a space or not, etc.

To make that easier, consider all values of `s` instead and concatenate
them at the last moment. That allows hook scripts to just add / remove
from the array as needed.