[FUN-10159] Add current execution context to the data model
Primarily intended for Stencils so that hook script can find out if
they're running on the Admin or Public side, however that may be a
better way to implement the `IsAdminUI` tag rather than looking at ports
in `global.cfg`.

To implement this I had to move the `ExecutionContext` class to
`publicui-core` so that it can be used in the transaction.

Aside: Removed the deprecated `Novell` execution context.