Make URL fill facets behave like other facets
URL Fill facets were different from other facets because the currently
selected values were not returned in the list of `CategoryValue`s.

For example when drilling down to `server.com/folder1/folder2/`, the
data model would only contain values below `folder2/` and would not
contain a selected value for `folder1/` nor `folder1/folder2/`. On the
other hand, A similar hierarchical metadata facet would contain selected
values for the parent categories (for example when selecting City =
Canberra on a Location facet, the data model would contain a value for
Country = Australia and State = ACT, both with `selected=true`).

This is partly due to the nature of their URL parameter, where a single
URL parameter is used to select multiple values, such as
`f.URL|url=folder1/folder2`, whereas typical hierarchical metadata
facets would use multiple parameters `f.Country|X=Autralia`,
`f.State|Y=ACT`.

This resulted in additional complexity around URL facets as the missing
values needing to be manually computed especially to build breadcrumbs +
some special cases in the templates to deal with the difference in URL
parameters. This in turns impacted the new data model for facets being
built for the Stencils (FUN-9578).

This changes this behaviour so that parent categories will now appear in
the data model, with `selected=true`. This is done by modifying the code
that generates `CategoryValue`s to yield a value for each parent, rather
than just for the folders just below the current one. With that change
URL facets are identical to other facet types.

I also had to modify the `Category` FreeMarker macro to remove the
special case to hide selected categories for URL facets, and replace
that with a more generic check to the `selected` flag on the value.

I also had to modify the tests accordingly, in fact some of them were
using nonsensical PADRE XML data because when a folder is selected (for
example `Shakespeare/clopatra/`) PADRE will not return counts for sibling
folders (`Shakespeare/romeo_juliet/`). This didn't show before as we
were only returning children of the selected category.

This will impact existing implementations as parent categories will now
be displayed, however:
- They will not be displayed if they use the default FreeMarker macros
as the macro has been fixed to address it
- The fix is easy enough to apply to custom macros (See the change on
`funnelback_classic.ftl`)
- A customary survey on all our pods show that only a few collections
use URL facets (A couple on the demo server + the intranet package
showcase - so apparently no production collection - I suspect URL facets
are used more in intranets / self-hosted customers)

So I think the change is acceptable (and I'll update the release notes
accordingly).