Add a customData map to the SearchQuestion
This is needed by the Facets Stencils, especially to make things work
with extra searches: The Stencils need to get the current query string
to build various URLs (for example the URL to select a facet).

This is obtained via Spring's `RequestContextHolder` but that is only
available in the main thread (so not for extra searches). By having a
custom map in the question the main search can obtain the query string,
store it in the question, and that will automatically get copied in the
extra searches that are generated (Via `SearchQuestionBinder`).

I will need to backport that to 15.10 and 15.8, as we want the Facets
Stencils to work with these versions.