Updated facet scope behavior to be similar to normal URL params.
FUNNELAPPS-27

When multiple URL parameters are set e.g. `x=1&x=2` this results in `x` being equal to the list of `1` and `2`. The `facteScope` does not follow this and instead overrides values of `x` e.g. `facetScope=x%3D3&x=1&x=2` results in x being set to `3` and values `1` and `2` are ignored.

This change could be potentially dangerous if someone was relying on the old behavior ie values in `facetScope` can override other values set outside of `facetScope`.