Support multi select/values from unscoped query facets.
FUN-7090

Ignoring the changes made to the model where we include the select/unslect URL. This adds support for any combination of:
* Category join type OR/AND. e.g. cars that are either red OR blue / houses that have both a bar and a pool. (Note that drill down SINGLE select wants to AND between the Categories)
* Select type SINGLE/MULTIPLE e.g. standard drill down facets or radio button facets/selecting multiple values like red or blue.
* Facet values are from SCOPED query or UNSCOPED query e.g. scoped is the normal query with facets values applied, makes sense for drilling down where unscoped is the traditional check box facets where you show all possible values.

Note that SINGLE is sort of weird and might be wrong for drill down facets because:
* when we drill down we end up selecting multiple values e.g. Earth -> Australia -> NSW -> Cowra
* when the metadata at a level is not a single value for each document e.g. a document is in both Earth and Mars. If we select Earth Mars will still be shown! Attempts have ben made to work around this FUN-4462 but those only worked for gscope facets and don't work for metadata based facets. We may need to do something about that.

Lots of this is stolen from the facet stencils note that:
* this fixes a bug in the stencils where SINGLE select would not unselect siblings which means gscopes probably don't work in the stencils version.

Counts are interesting. Generally if the Selection is OR counts will not be displayed as we don't have a way to work these out. To account for this the count type has been changed from a int to Integer and can be nulled. Templates using he OR type (none of these exist) will need to account for this. When in the AND mode and values are from the unscoped query if the value is only in the unscoped query then we know that the count is zero so the AND case should work when in MULTIPLE select mode.