Add All document facet category.
FUN-10459

This makes it easy to add a ALL tab facet.

I have also changed the CategoryDefinition to ensure that the facet URL param key is consistent it enforces that they start with `f.<facet name>|`.

I also realised a bug in that MultiFacetedNavigation was calculating counts for all facet values when it was only intended to do that for unselect ones. I have left a TODO and will update this in another MR.