Improve how we work out how much time was spent executing the query.
FUN-10459

Before we counted either the time we waited or how long the query took from metrics in the search transacton, we took the max value of those. This might not make sense as we might first wait for a query which is not the first query to finish which means we count more time for that query than it actually took. Now we prefer to use the time spent in the search transaction rather than the time we waited.

Also add the unit tests that where not added to the previous MR

Also improve the logging in MultiFacetedNavigation so we can see how many extra searchers where executed.