Limit the number of concurrent extra searches run by a single search.
FUN-10592

The global TaskExecutor will execute up to 100 threads at a time, if a single submits 1000 extra searches then we will have running 100 threads this can make the server unresponsive.

To avoid this a LimitedCPUUsageExecutorHelper which limits the number of executing tasks a single search can have. By default it limits the number running extra searches to the number of available processors.

This can be configured using
```
ui.modern.extra_search_cpu_count_percentage=100
```
if set to `50` then at most the number of extra searches that can be executing at once is `50%` of the available processors. e.g. on a 4 core machine only 2 extra searches will run at a time.

This value rounds down so to get one less than the available processors you can set this to `99`.

This also supports overallocation as well.

The `LimitedCPUUsageExecutorHelper` uses a semaphore that is local to the executing search and will not block other searches from running. Should a problem exist with the semaphore other searches likely wont be affected.