Add support for getting all values
FUN-10459

In tabs if the search returns no results or results from only some category values we still want to show ALL possible values.

This change makes it possible. In the case the category definitions where the user defined the values like gscope or clive (collection) we have the category just add the known category value to the list of values. This does not require an extra search. In the case the category definitions get their values from the index like metadata or dates we run a `!padrenull` query to find those values. Running a query rather than doing pre-caching means this will work in DLS at the cost of more CPU cycles per query.