Use the running time of the Callable executed.
FUN-10592

As this gives us a much more accurate time for how long each extra search took. Also because the callable updates the total time taken in extra searches, we avoid a bug where:
* if ui.modern.extra_search_cpu_count_percentage results in less than the number of searches we want to execute.
* we would submit all most all extra searches because we didn't update how much time we had left until after we submitted and possibly ran most extra searches.
As we update the time used as soon as the extra search is done, we can take this into account when submitting extra searches and stopp submitting when we run out of time.