Ensure facet optimisations are applied.
FUN-7545

This adds some facet optomisations and I make a change to the search life cycle to add a way to ensure some query processor options are set. We want to do this in the case we are running the query to work out the counts for each category value.