Add support for Query based facets.
FUN-10716

When reading Query category facet config from the index then the gscope number is the one set in the xml file (as that is the gscope number used when the index was built). When reading query categories from config we generate the gscope name from the query, in that case we also convert the `QueryCategory` to a `GscopeItem`, as it is implemented with gscopes.