Ensure selected values are in the facet model.
FUN-10798

This change makes it so that selected values are always in the facet model. If the search is changed while facets stay selected it was possible to have the constraint remain but no facet values shown so it could not be removed. In the case of unselecting OR values this would lead to a state that could not be left. This change ensures that the selected category values are always in the model so we can always unselect those values.

This also fixes URL Fill (URL drill down) facets to work with `FROM_SCOPED_QUERY_HIDE_UNSELECTED_PARENT_VALUES`.

This also changes URL Fill (URL drill down) facets such that selected values are nested like other facet types.