Returns XML documents as XML and not parsed via jsoup as HTML
FUN-9700

A issue existed in the cache copy controller where documents that where raw byte records (e.g. form push/web) where never being treated as XML even if the content type or filtered content type was XML. This change uses the new `ConvertRecordType` class to convert the raw byte record to a XML one if it is in-fact a XML document. This results in the record not being parsed by jsoup resulting in the XML elements not being lower cased.

Note that the XMLUtils always return UTF-8 XML and some how the modern UI is also setting the content type HTTP headers to XML with UTF-8 charset.