This fixes a bug where counts are wrong for collection fill in tabs.
FUN-10656

The issue is when the all query is not run (because this is a user defined category so the all query is not needed), we use the standard query to get the collections to get counts from. This results in us only getting a count for the selected collection. This fixes that by in values are from all query mode we look for the counts of all collections in then collection fill category.