Increase the padre timeout for service users.
FUN-10969

Often we need to run padre-sw during indexing time to find out something about the index we have made. This is done as offline task as we expect it to be long. This can take longer than the default 30s of timeout which is allowed by default. This change makes it so that when a service user (the user that makes AA or CA padre-sw requests) runs a query the timeout is increased to 20mins. This should make it so that updates work out of the box without needing the globally increase the padre timeout.