Prefer the lz4 frame output stream as it is faster for both compression and decompression.
for compression of 90MB it went from 900ms to 800ms, for decompression it went from 34ms to 24ms.

This is only used for linux as we don't want to use the native compressors for windows. I bother to add this because simon is so concerned about CPU usage, although it probably make no difference.