Adds sum support for groovy scripts that put the wrong type into the gscope count map.
The type for the `gscoperCount` has changed from `Map<Integer,String>` to `Map<String, Integer>`. What I have done is wrapped `gscoperCount` map such that if you try to use a `Integer` key it will convert it to a `String` also if the wrong type is used a warning is given. If debug is on a stack trace is printed this may be needed to find where the problem is.