Don't require gscopes have a count greater than zero
SUPPORT-2565

I suspect the reason this came in is because at some point I had a bug in padre where it was returning zero counts for gscopes, I later removed this when I realised this was not the previous behavior of padre and it is not inline with how metadata works. Padre at the moment will never return gscope counts with a zero count, so under normal operation this check would never be useful. It is un-helpful for older implementations which inject zero value gscopes to trick the modern UI to create a facet category for the gscope.