Fixes a issue where falling back to the referrer URL may not work
FUN-11146

if the parameter we are interested in is the first e.g. if click tracking controller is not passed the query (which can only happen if the click tracking URL is changed) and the referrer URL looks like:
```
https://foo.com/bar?query=foo&collection=bar
```
This is because the query string methods (like toSingleMap) expect a query string not a entire query.