Allow for javascript in javadoc comments - New option in JDK 8u121 apprently
It seems that without this, the javadoc build is aborted before completion, meaning we get broken links in the docs collection.

Note: This will break the build in any earlier java version (facepalm) as it's an unknown option there

I looked for a little to see if I could find what script is used in a comment to see if it could be removed
instead, but I couldn't find what it was complaining about (and unfortunatly the error message it gives isn't
very helpful). In any case, I think we can trust our javadoc commetns to be a safe source of content, so if we
wanted to add javadoc to them, we should probably be allowed to.